Fix `select` builder test (#3853)
### Ticket
Closes #2488 
Closes #2496 

### Problem description
`select` was not passing in the `stride` argument due to #2488. This was
causing the floating point exception seen in #2496.

### What's changed
Re-enabled passing the `stride` argument in, and some previous change
fixed the issue seen in #2488. This argument fixes the floating point
exception seen in #2496, and brings `select` golden testing into the CI

### Checklist
- [x] New/Existing tests provide coverage for changes