Add non-blocking toHost, implement wait in ttnn runtime (#3872)
### Ticket
Part of #3698, closes #3176

### Problem description
Currently `toHost` always blocks on device completion.

### What's changed
Added an option to make `toHost` non-blocking, and implemented `wait`
for the user to explicitly block on device completion. This is a
prerequisite for multi cq workloads in the future. Without this we won't
be able to reap any perf benefits from multi cq optimizations.

#### Misc
* Removed raw host data pointer workaround as with the new uplift metal
has a generic API of getting the host buffer
* Removed unused data structures in utils
* Added mnist_linear trace test

### Checklist
- [X] New/Existing tests provide coverage for changes