Add `ttmlir-reduce` to build process
This change adds a simple customization to `mlir-reduce` (see
https://mlir.llvm.org/docs/Tools/mlir-reduce/)" so that it functions
with TT dialects. This is the simplest implementation, and will use
standard reduction techniques to reduce an input while maintaining a
user defined "interestingness". Useful for minimizing an input program
such that it still exhibits some behavior. Notably should be used to
create a minimal program to reproduce some error.