Support blocking over dst (#3886)
This commit changes the OptimizeTensorLayout pass to additionally
constrain layouts by programming
the generic op's `block_factors` such that resulting block shapes never
exceed a volume that fits into
the destination register. These semantics are subject to change in the
future, but they are sufficient for
now in order to unblock matmul performance sweeps.

---------

Co-authored-by: Vincent Wells <vwells@tenstorrent.com>