Add unique locs pass (#3360)
### Ticket
Closes #3336

### Problem description
A pass should be added to `ttir-to-ttnn-backend-pipeline` to check if
all ops have a unique `NameLoc`.
Op overrides (Conv2dConfig, OutputLayout, ...) are applied to ops using
their `NameLoc`. If a ttir op is decomposed to multiple ttnn ops, they
will initialy share the same location. This pass will enforce adding a
suffix to locs of newly generated ttnn ops, so that the overrides are
applied only to the targeted op.

E.g. `ttir.conv2d` generates `ttnn.prepare_conv2d_weights`,
`ttnn.covn2d` and `ttnn.reshape`. Only `ttnn.conv2d` should have the
same location (and overrides) as `ttir.conv2d`, while the locs of other
ops should have an appropriate suffix added.

### What's changed
Added a new pass called `TTNNUniqueLocations` to check if name locations
of ttnn ops are unique. If the condition is not met, it will emit errors
for all ops that have non-unique name locs and signal pass failure.

Added suffix to generated ops.

**UPDATE:** This pass will be **disabled** by default. Although suffixes have been 
added so that ops relevant to the optimizer have unique name locations, enabling 
this pass by default would potentially cause a lot of headache in the future.
It should be enabled when we want to ensure that overrides are applied appropriately
or perf data for individual ops is correct.