Add fusing path to match/rewrite scatter op intoupdate/fill_cache (#3676)
### Ticket
We are looking to [support generative
models](https://github.com/tenstorrent/tt-torch/issues/614)

### Problem description
To support generative models we need to enable KV-static cache
fill/updates. Once we lower the module to stableHLO the cache update is
a 20+ op pattern that start with a scatter op and can be traced all the
way to a tensor containing the cache indices that need to be updated. We
want to add a pattern matcher for scatter ops that represent a kv-cache
fill/update and replace with a single Fill/UpdateCacheOp.

### What's changed
- Add a pattern matcher to TTIRFusing pass to determine if a scatter op
can be replaced with a FillCacheOp or UpdateCacheOp
- Move conditions from ttir::ScatterOp::verify() into TTIRtoTTNN pass to
allow kv-cache scattter ops to be rewritten into Cache ops
- Added test cases for FillCache and UpdateCache rewrites.

### Checklist
- [x] New/Existing tests provide coverage for changes