Make ttir to ttir decomp configurable (#3908)
### Ticket
NA

### Problem description
TTIRToTTIRDecomp is a bit over-tailored to TTNN's use case. For some CPU
fallback work I'm doing, customizing which ops are actually decomposed
is useful; in the future, it is likely that we might want some patterns
only for TTNN and not TTMetal or vice versa as well.

### What's changed
TTIRToTTIRDecomp pass now can optionally take a set of specific ops to
decompose; if the option is not set, current behavior remains.

### Checklist
- [X] New/Existing tests provide coverage for changes