TTIR to TOSA Hoist Infra Changes (#3910)
### Ticket
NA

### Problem description
Lowering some of the ops we need from TTIR to Linalg for CPU fallback
requires bouncing through TOSA. This requires some minor infra changes
to the hoist process, because TOSA ops are not DPS but we need to
preserve the relationship. In an effort to make PRs more readable, I
have separate the related infra changes + transforms into a standalone
PR, which will be exercised in follow-on PRs which actually add the TOSA
conversion patterns and testing.

### What's changed
Hoist mechanics change to use temporary return value, store which param
represents DPS output as an attr, and use a pair of workarounds to
restore the output to the "output" param after TOSA conversion if needed
+ prune return values after they've been used s.t. we don't get unwanted
RVOs.

### Checklist
- [X] New/Existing tests provide coverage for changes