Canonicalize all transpose ops to permute (#3873)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3877

### Problem description
Transpose and PermuteOp can perform the same function, with the caveat
that PermuteOp can perform multiple transposes at once. Supporting both
these ops throughout all passes in the TTIRToTTNNPipeline results in
much nearly identical logic.

### What's changed
- Canonicalize TransposeOp to PermuteOp in ttir
- Remove all TransposeOp handling logic
- Remove all tests pertaining to transpose specifically, change file
check tests to expect permute instead.
- **We will leave the transpose op in the TTNN dialect as it may make
for a more user-friendly output when using the emitc path. We can choose
to split `ttnn.permute` into `ttnn.transpose` in the ttnn dialect in the
future as the canonicalization is not registered in that dialect**

### Checklist
- [x] New/Existing tests provide coverage for changes