Adding eltwise binary output dtype argument in TTNN dialect (#3934)
### Ticket
Closes https://github.com/tenstorrent/tt-mlir/issues/3921

### Problem description
We currently lack compiler support for specifying the output data type
of element-wise binary operations. The TTNN runtime models this as an
output based on the result type. However, the generated C++ code does
not handle this similarly, leading to discrepancies.

### What's changed
Added optional output dtype parameter to eltwise binary ops in TTNN
dialect.

### Checklist
- [x] New/Existing tests provide coverage for changes