[Optimizer] Insert row major workarounds  (#3953)
Adds a change in Optimizer that goes through ops and for specific ops
(Upsample MaxPool2d currently) we check if they satisfy constraints in
current layout. If not, we try converting them to RowMajor. If that
fails, we fail the compilation.

However, MaxPool2d constraints api is broken in tt-metal
(https://github.com/tenstorrent/tt-metal/issues/24358) so for that op we
convert to RowMajor always.

Potentially we can turn this on for other ops too.