Support `UpsampleOp` in `bilinear` mode (#3978)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3396

### Problem description
In the general case, `UpsampleOp` would assert in `tt-metal` on the
requirement that input has to be in height-sharded configuration when
the mode is `bilinear`.

### What's changed
Implemented a workaround to height shard the input when the mode is
`bilinear`. It was implemented as a three-stage rewrite pattern where:
1. Input has to be in `RowMajor`/`BFloat16`. This is irregardless of the
mode. Even though this pattern already exists as part of layout
workarounds, as layout workarounds are applied after decomposition
workarounds I had to rewrite it once again here.
2. Channel axis size has to be a multiple of the tile width. Added
padding to the channel axis to the nearest multiple of the tile width,
after which we have to return the size to the original value (using
`SliceOp`).
3. Finally after previous stages, height shard the input.

### Checklist
- [x] New/Existing tests provide coverage for changes