fixing const-eval with quantize (quantize ttnn -> flatbuffer) (#3956)
### Ticket
Fixes #3505 

The scale and zero point for quantization are inside constants and
should be extracted as scalars in the per-tensor case. As such we need
to enter into the const-eval function and go through the const-eval
chain (walk through typecast (optional) -> to_device -> to_layout ->
from_device) to reach the scalar value. This PR adds the logic for
entering the const-eval function and continuing with the search.