Erase Inverse Ops upgrades (#3932)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/3877 - Part 3

### Problem description
In many cases, further reductions in the number of TMs in a graph
require commuting downwards as well, EIO currently does not do this.

### What's changed
- Added commute down logic. Continually commute TMs upwards, then
downwards, until commuting in either direction erases no ops.
- TODO: No implementation for commuting reshape downward through
broadcast
- TODO: No implementation for commuting permute downward through
broadcast
- Added commute pattern for ConcatOp
- TODO: No implementation for commuting reshape downward through concat

The TODOs listed will not cause the pass to fail, their respective
`isCommute<DIRECTION>Viable` methods return false and so those
unimplemented functions are not executed.

**This PR contains the changes necessary to achieve a minimal-TM graph
for resnet**, and then some.

CC: @odjuricicTT @rpavlovicTT @svuckovicTT @sdjordjevicTT @mtopalovicTT 

### Checklist
- [X] New/Existing tests provide coverage for changes

---------

Co-authored-by: Ognjen Djuricic <odjuricic@tenstorrent.com>