Add metal override option to PR (#4018)
### Ticket
slack

### Problem description
During validation of tt-metal commits in an uplift against tt-mlir it
would be nice to have a tt_metal override option in the onPR.
When running tt-metal -> tt-mlir uplift tests, tt-mlir build times are
consistently extremely slow at around 1hr, as it uses ubuntu-latest
runners and caches are reset.

### What's changed
Added option to dispatch onPR with tt-metal override sha.
Added option to run onPR build on builder machine which is much faster
(but there are only 2 of them available).
Moved wheel building to ubuntu-latest (it was failing)

### Checklist
- [ ] New/Existing tests provide coverage for changes