Add a pass to legalize StableHLO composite operations directly to their TTIR counterparts (#3881)
### Problem description
On the frontend side, some operations get decomposed to many StableHLO
operations. Intended way to deal with that is to wrap an operation in a
StableHLO CompositeOp, so that we can avoid having to write brittle
fusion passes.

### What's changed
A new pass was added to legalize CompositeOps directly to corresponding
TTIR ops.

### Checklist
- [X] New/Existing tests provide coverage for changes