Check for deleted files as well in the CI (#4035)
### Ticket
/

### Problem description
`Ignore files` job in the CI was only checking if any files other than
`./docs/*` have been **changed**, not including **deleted** ones in that
list.
That caused the `On PR` workflow to be skipped if your last commit only
removed files.

### What's changed
We are now checking for deleted files as well when checking whether to
run the `On PR` workflow.
We are now outputting the list of deleted files and checking deletions
(in addition to changes) in files. The check is happening in the
`build-image` job.

### Checklist
- [ ] New/Existing tests provide coverage for changes