Grid shape verification in `TTNNLayoutAttr` (#4016)
### Ticket
https://github.com/tenstorrent/tt-mlir/issues/4012

### Problem description
Currently, `gridAttr` is unrestricted when buffer type is not L1 (system
memory or DRAM). This leads to various cases where TTNNLayoutAttrs are
effectively the same, but due to different grid sizes it's much harder
to compare them. As grid size is not a factor when buffer type is not L1
we can opt to always use {1, 1} as default grid size.

### What's changed
Added a verification that verifies that grid shape is `1x1` whenever
buffer type is not-L1.

### Checklist
- [x] New/Existing tests provide coverage for changes