Add fusion pattern to fuse PadOp into PoolingOp padding attribute (#4055)
### Ticket
N/A

### Problem description
We may receive an IR which explicitly places padding for a pooling op on
the pooling input. This can make it impossible to commute the reshape +
permute that is inserted on the input of the op during
`TTIRFlattenSlidingWindow`. Also, this is unnecessary.

### What's changed
Add a pattern to `TTIRFusing` to fuse PadOp into PoolingOp

### Checklist
- [x] New/Existing tests provide coverage for changes

---------

Co-authored-by: Aleksandar Zecevic <azecevic@tenstorrent.com>