Rework TTIR ViewLayoutOp to include affine map directly + remove workaround (#4049)
### Ticket
[Link to Github Issue
](https://github.com/tenstorrent/tt-mlir/issues/3812)

### Problem description
As part of MetalLayout refactoring, we want to cleanup ViewLayoutOp s.t.
we don't need a workaround to compute the affine map at the last minute
before usage; instead, when the op is created we should compute + store
the affine map it represents.

### What's changed
ViewLayoutOp stores its own affine map, which is computed on
construction. This is later used when the view is applied, eliminating a
clunky workaround. Various testcases are updated s.t. their view_layouts
are correct. ViewLayoutOp's verifier can now verify the input maps to
the output shape via given affine map.

### Checklist
- [X] New/Existing tests provide coverage for changes