Run builder tests on `p150`s (blackholes) in CI (#3913)
### Ticket
Closes #3898

### Problem description
We are not running builder tests on blackholes in CI

### What's changed
- Adds builder tests to CI on blackhole machines (specifically `p150`s)
- Adds clarity to a `ttrt run` flag, elucidating its necessity on
blackholes
- Adds a set based skip mechanism for tests that fail to function on
specific intersections of targets and systems. See the `skip_config`
mark in `pytest.ini` for more information.

### Notes
Also fixes alignment in `build-and-test.yml`. This makes the diff a
little hard to read, the only material change is on line 177 where a new
matrix entry is added for `p150`

### Future Work
Once more blackhole machines (`p300`, etc) are added to the CI, these
must also be added to the `build-and-test.yml` matrix

### Checklist
- [x] New/Existing tests provide coverage for changes