Fuse average pooling when divisor is a pool of a constant (#4060)
### Ticket
N/A

### Problem description
torch-xla will provide an IR for average pool which consists of two
`ttir.pooling` ops. One sum pool on the activations, and one sum pool on
a constant tensor. Then, divide the first by the second.

The divisor pooling op is not only an unnecessary computation, but is
not a valid pooling operation in either torch or ttnn. This is because
the input has no channels or batch, it has only the spatial dims. So, we
could not rely on consteval to handle this.

### What's changed
Added a fusing pattern to pick up on the pattern `div(sum_pooling(act),
sum_pooling(const))` and replace it with `avg_pooling(act)`

### Checklist
- [X] New/Existing tests provide coverage for changes