Run workarounds pass for full op when optimizer enabled (#4149)
### Ticket
Related to #3912 

### Problem description
I was hitting the same data type mismatch issue as #3912 when bringing
up a resnet demo on tt-torch with optimizer enabled. Turns out it was
because the data type workaround for full op is not applied.

### What's changed
Apply workarounds pass for full op when optimizer is enabled.
FYI @ddilbazTT @sgholamiTT 

### Checklist
- [X] New/Existing tests provide coverage for changes