Skip creation ops for trace capture, update conv verify logic, don't allocate empty slots for constants/params (#4165)
### Problem description
We should be skipping creation ops for trace since they're not
supported. Conv verification was failing when conv ops are hoisted into
trace functions if their weights are being prepared with
`prepare_conv2d_weights` op. We were allocating device buffers for
constants/params (Thanks @svuckovicTT for catching this).

### What's changed
* Creation ops are now skipped in the trace hoist transform pass.
* When within a trace function, conv verifier will follow the function
argument to the original location where the input is generated, and run
the verification logic on the actual input instead of the function
parameter.
* Only allocate separate buffers on device for inputs, for constants and
parameters, use the original tensor directly.
### Checklist
- [X] New/Existing tests provide coverage for changes