Fix prepare dtype usage (#4335)
`output_dtype` attribute on prepare conv/bias should be used to indicate
to prepare api what is the output dtype of conv operation itself, wheres
our code uses it to specify the output dtype of prepare operation.
To tell ttnn to which dtype output of prepare should be we need to use
`weights_dtype`.
 
This fix sets `weights_dtype` on prepare apis to the same type as input
so that after prepare bias, weight and input are in same dtype.
`input_dtype/output_dtype` on prepare APIs is updated to dtype of input
tensor and dtype of conv result respectively.

fixes #4180