[Utils] Updated generateReshape to accept location parameter (#4422)
### Problem description
When creating a reshape function using `generateReshape` from
`lib/Conversion/TTIRToTTNN/Utils.cpp`, the newly created reshape will
always inherit the location from the input. This is not always desirable
because sometimes we want to reshape the input to a certain operation
and inherit the location from that operation, not the input.

This posed a problem when using the Chisle tool for golden verification
because Chisle uses locations to track the relationship between ops in
the TTNN dialect and the TTIR dialect from which they originated, for
which we have custom golden implementations. The problem is that if a
certain operation’s input is a function argument, the generated reshape
will inherit the location of the input rather than the operation, and
thus the Chisle tool will not know that the generated reshape originated
from the operation.

### What's changed
Updated the function to accept a location and changed its usage to
adhere to the new signature.

### Checklist
- [x] New/Existing tests provide coverage for changes