[optimizer] allow passing external device
Optimizer needs an open device to work - this causes issues in frontends
where we need to close the device before running the `ttir-to-ttnn`
pipeline and then re-open the device afterwards. New field is added in
the pipeline options, allowing us to pass a device pointer to the
optimizer. If the frontend doesn't pass the device pointer, the
optimizer will (try to) open its own device.

The added field is not an `Option<T>`, since we do not want
to parse it - for example when running the pipeline through the CLI.
This means that, for now, we will not be able to use this mechanism in
`tt-forge-fe` - due to the RTTI we cannot use the pipeline options
struct directly. We'll need to figure out how to workaround this.

The `SingletonDeviceContext` class is refactored to allow for this.
Additionally, the behaviour of the singleton is changed so that the
lieftime of the device under it is managed fully by the users of the
singleton. Even before this change, the situation with the singleton was
a bit complex, since we allowed users to close the device manually, but
opening of the device was done only during construction of the
singleton. This caused issues if you wanted to run optimizer multiple
times in a single process (#4232).

Asserts are in place to verify that the singleton is used properly. More
info on this in the code comments.

Needed For tenstorrent/tt-xla#952
Closes #4232