[Chisel] Various bugfixes (#4437)
### Problem description
There were various problems occurring when running Chisel.

### What's changed
- Added missing corresponding golden operations for `ttir.full`,
`ttir.sign`, `ttir.clamp_scalar`, `ttir.clamp_tensor`
- Cast the output of the golden operators to their input dtype to handle
cases where, for example, a `ttir.subtract` operation had a `ttir.gt `as
its RHS, which returned a boolean instead of f32.
Added support for generating random inputs when the tensor dtype is an
integer
- When populating input tensors from file paths, the file paths are now
sorted. This is not ideal, and in the future we should support naming
files to correspond to specific inputs

---------

Co-authored-by: Nikola Drakulic <ndrakulic@tenstorrent.com>