Add conversion from `stablehlo.custom_call(@mhlo.erf)` to `ttir.erf`, add `gelu` fusion pattern (#4442)
### Ticket
N/a

### Problem description
- torch-xla will provide calls to `erf` in the form
`stablehlo.custom_call(@mhlo.erf)`. We do not lower this yet
- Gelu is provided in a multitude of forms. These are some eltwise
operations surrounding a call to `erf` or `tanh`

### What's changed
- Lower `stablehlo.custom_call(@mhlo.erf)` to `ttir.erf`
- Add fusion pattern, capable of fusing a few different forms of `gelu`
as they are presented from torch-xla and jax
- Add test for SHLO lowering
- Add tests for `gelu` fusion.

**NOTE**: The `erf` based version of `gelu` as it is provided from jax
lowers `erf` as a polynomial approximation. Unlike torch-xla which
conveniently provides `stablehlo.custom_call(@mhlo.erf)`. tt-xla
currently has a monkeypatch to return `gelu` as a composite call, which
is lowered properly, thereby bypassing this issue. So, the fusion of the
erf polynomial isn't urgent and may not even be necessary.

### Checklist
- [X] New/Existing tests provide coverage for changes