Docs building incorrect docker image pull (#4503)
### Ticket
/

### Problem description
Docs building and on-push get triggered at the same time, that
potentially causes an issue where docs build pulls the "latest" docker
image before the on-push has tagged the newest image as "latest".
Practically docs build then uses an older "latest" docker image, not the
one from the commit it was triggered on.

### What's changed
The docs build & deploy is not triggered at the end of the on-push
workflow, that gets rids of the race condition issue with the "latest"
docker image.

### Checklist
- [ ] New/Existing tests provide coverage for changes