[Optimizer] Add Constraint API for CNN related ops (#4436)
### Ticket
#4204 

### Problem description
We need constraint API for other CNN related ops.

### What's changed
This PR:
- Add **constraint** API for `TTNN_PrepareConv2dWeightsOp`,
`TTNN_PrepareConv2dBiasOp`, `TTNN_AvgPool2dOp`, `TTNN_BatchNormOp`.
- Add **runtime** API for `TTNN_AvgPool2dOp`, `TTNN_BatchNormOp` (the
other two ops have memory effects, so cannot have the runTime API. Read
more
[here](https://github.com/tenstorrent/tt-mlir/issues/4199#issuecomment-3140045496)).
- Add UTs.
-Enables tests for `TTNN_MaxPool2dOp` (which was originally disabled due
to issue 2976), but there are some workarounds that could be done for it
to enable the tests).
- Just for the record; This PR does NOT resolve the issue, it just finds
workarounds to bypass it for tests related to `maxPool2DOp`.


### Checklist
- [X] New/Existing tests provide coverage for changes