Reorganize full model tests (#4526)
### Ticket
N/A

### Problem description
There are various model tests that appear in different contexts (i.e.
different components are running them as E2E tests). They are de facto
artifacts of frontends, so once they get outdated, we have to update
them in multiple places (and sometimes we end up with slightly different
versions of models depending on fact if the change is breaking or not).

### What's changed
Created a common directory `test/ttmlir/models` where such TTIR files
reside. These files are not run as part of `llvm-lit` tests; they only
serve as input files to other components. With this approach, there is
only a single source of TTIR for each model, which facilitates
maintenance and mitigates the risk of running different versions in
different components.

### Checklist
- [x] New/Existing tests provide coverage for changes