Add Test Metadata to JUnit XML  (#4407)
### Ticket
Part of #3899 

### Problem description
We need more detailed information in the test reports coming out of the
`pytest`s for `ttir-builder`. This information includes but is not
limited to input shape size, input dtypes, failure stages, op specific
parameters/attributes. etc.

### What's changed
A hook has been added to `conftest.py` that gathers all the information
noted above, formatting it for easy ingestion by whatever wants the read
the produced JUnit XML. A thing to note is that since we now keep track
of specific failure stages, any test that would need to be skipped for
fear of failing a CI run can now instead be XFAILed, allowing us to see
more information about the tests that fail.

### What still needs to be done
This PR is part of a larger effort to ingest the results of this testing
data into superset for easy viewing and triaging. This effort takes
place over a few different repos, namely this one, `tt-github-actions`,
and the airflow repo. Finally, once the parsers and schemas in the other
repos are merged in, then the CI itself must be updated to correctly
link up all this information into a pipeline

Secondly, the framework for the runtime API being called by `pytest` in
order to execute the generated flatbuffers in a single test process
(with children) has been added and sits dormant until that is added.
This will come after the superset integration, and _should_ start being
reflected in superset seamlessly with this change.

### Checklist
- [x] New/Existing tests provide coverage for changes