Merge Conv2dConfig structures (#4515)
### Ticket
Closes #3452

### Problem description

Currently we have multiple classes that store Conv2dConfig parameters:
- `Conv2dConfigAttr`
- `Conv2dConfigAttrParams`
- `Conv2dConfigOverrideParams`

First one is an immutable MLIR attribute and should not be changed.
Second one is a helper class used to overcome the immutability of the
MLIR attribute when setting individual parameters.
Third one is a class for storing and parsing Conv2dConfig overrides.

Functionalities of second and third class can be merged into a single
class.

### What's changed
Merged `Conv2dConfigAttrParams` and `Conv2dConfigOverrideParams` classes
into a single class called `Conv2dConfigParams` that has the
functionality of both classes.